requirements.txt

numpy
scipy
pandas
matplotlib
plum-dispatch